Move scripts to generators folder
