+++ scripts/generators/generate_chains_list/generate_chains_list.py
